Current Climate Data for Variables of Interest
The added files are the current global climate data from WorldClim at a resolution of 5. The wc2.1_5m_bio_2.tif file corresponds to the mean diurnal range data, the wc2.1_5m_bio_5.tif file corresponds to the maximum temperature (degrees Celsius) of the warmest month, the wc2.1_5m_bio_11.tif file corresponds to the average temperature of the coldest quarter, and the wc2.1_5m_bio_14.tif file corresponds to the precipitation data of the driest month.